refactor: remove 'format' vs 'lint' differentiation
unifies into one name, 'fix'. the differentiation was done in the first
place because formatters typically need less configuration and don't
output errors. but if that is so, we can include them with 'lint' since
they won't output errors. and since formatters aren't linters, we can
create a new 'fix' name